Add files via upload
Case_control_data.tsv is a version of BRCA_BRIDGES_case_control_clinical.csv converted to TSV and with the addition of a first column meant to be an ID.  In this first column, the header / label is "Case_control_data", and in subsequent rows, the ID is 1 to N ascending.  When uploaded as a Terra data table, this will produce a table named "Case_control_data".
Uncalibrated_functional_assay.tsv  is a version of BRCA1_RING_BRCT_FA_sge_scores.csv likewise processed: converted to TSV, and with the addition of a first column that has the header label "Uncalibrated_functional_assay" and with 1 to N ascending in subsequent rows.